+++ src/term.c
@@ -439,6 +439,7 @@ void
 
   if (full) {
     term.selected = false;
+    term.selection_eq_clipboard = false;
     term.hovering = false;
     term.hoverlink = -1;
     term.on_alt_screen = false;
@@ -3435,8 +3436,9 @@ void
               d[1].attr.attr |= TATTR_OVERHANG;
             }
 
-            d->attr.attr &= ~ATTR_FGMASK;
-            d->attr.attr |= TATTR_EMOJI | e.len;
+            // emoji length is encoded in font family
+            d->attr.attr &= ~FONTFAM_MASK;
+            d->attr.attr |= TATTR_EMOJI | ((cattrflags)e.len << ATTR_FONTFAM_SHIFT);
 
             //d->attr.truefg = (uint)e;
             struct emoji * ee = &e;
@@ -3493,6 +3495,7 @@ void
       scrpos.x = backward ? backward[j] : j;
       wchar tchar = d->chr;
       cattr tattr = d->attr;
+      //printf("%2d:%2d: %02X %08llX\n", i, j, tchar, tattr.attr);
 
      /* Many Windows fonts don't have the Unicode hyphen, but groff
       * uses it for man pages, so display it as the ASCII version.
@@ -3528,6 +3531,17 @@ void
 
         colour bg = win_get_colour(SEL_COLOUR_I);
         if (bg != (colour)-1) {
+          // enable ATTR_DIM to be applied to the effective background:
+          // (we could exchange assigned background and foreground values 
+          //  and exchange them back with ATTR_REVERSE,
+          //  but that would mangle emoji highlighting)
+          // we dim bg explicitly
+          if (cfg.selection_mode > 1) {
+            if (!term.selection_eq_clipboard)
+              // dim if selected && not in clipboard
+              bg = ((bg & 0xFEFEFEFE) >> 1) + ((win_get_colour(BG_COLOUR_I) & 0xFEFEFEFE) >> 1);
+          }
+
           tattr.truebg = bg;
           tattr.attr = (tattr.attr & ~ATTR_BGMASK) | (TRUE_COLOUR << ATTR_BGSHIFT);
 
@@ -3547,8 +3561,17 @@ void
             tattr.attr = (tattr.attr & ~ATTR_FGMASK) | (TRUE_COLOUR << ATTR_FGSHIFT);
           }
         }
-        else
+        else {
           tattr.attr ^= ATTR_REVERSE;
+          if (cfg.selection_mode > 1) {
+            if (!term.selection_eq_clipboard)
+              // dim if selected && not in clipboard
+              tattr.attr |= ATTR_DIM;
+            else
+              // work around old/new state glitch
+              tattr.attr &= ~ATTR_DIM;
+          }
+        }
       }
 
       if (term.hovering &&
@@ -3693,8 +3716,9 @@ void
               d[1].attr.attr |= TATTR_OVERHANG;
             }
 
-            d->attr.attr &= ~ATTR_FGMASK;
-            d->attr.attr |= TATTR_EMOJI | e.len;
+            // emoji length is encoded in font family
+            d->attr.attr &= ~FONTFAM_MASK;
+            d->attr.attr |= TATTR_EMOJI | ((cattrflags)e.len << ATTR_FONTFAM_SHIFT);
 
             //d->attr.truefg = (uint)e;
             struct emoji * ee = &e;
@@ -3906,7 +3930,11 @@ void
               disp = cfg.disp_tab;
             }
           }
-          tattr.attr &= ~(ATTR_BOLD | ATTR_DIM);
+          if (selected && cfg.selection_mode > 1)
+            // do not spoil dimmed selection highlighting
+            tattr.attr &= ~ATTR_BOLD;
+          else
+            tattr.attr &= ~(ATTR_BOLD | ATTR_DIM);
 
           if (!(cfg.disp_clear & 8))
             tattr.attr &= ~TATTR_CLEAR;
@@ -4244,7 +4272,6 @@ void
 #if defined(debug_dirty) && debug_dirty > 1
     printf("dirty ini %d:* lin %d run %d\n", i, dirty_line, dirty_run);
 #endif
-    cattr attr = CATTR_DEFAULT;
     int start = 0;
 
     displine->lattr = line->lattr;
@@ -4293,9 +4320,12 @@ void
       }
 #endif
       if (attr.attr & TATTR_EMOJI) {
-        int elen = attr.attr & ATTR_FGMASK;
+        // emoji length is encoded in font family
+        int elen = (attr.attr & FONTFAM_MASK) >> ATTR_FONTFAM_SHIFT;
+
         cattr eattr = attr;
         eattr.attr &= ~(TATTR_WIDE | TATTR_COMBINING);
+
         wchar esp[] = W("        ");
         if (elen) {
           if (!overlaying) {
@@ -4316,6 +4346,13 @@ void
                 bg = eattr.attr & ATTR_REVERSE
                           ? win_get_colour(BG_COLOUR_I)
                           : win_get_colour(FG_COLOUR_I);
+
+              if (cfg.selection_mode > 1) {
+                if (!term.selection_eq_clipboard)
+                  // dim if selected && not in clipboard
+                  bg = ((bg & 0xFEFEFEFE) >> 1) + ((win_get_colour(BG_COLOUR_I) & 0xFEFEFEFE) >> 1);
+              }
+
               eattr.truebg = bg;
               eattr.attr = (eattr.attr & ~ATTR_BGMASK) | (TRUE_COLOUR << ATTR_BGSHIFT);
               eattr.attr &= ~ATTR_REVERSE;
@@ -4413,6 +4450,7 @@ void
    /*
     * Third loop, for actual drawing.
     */
+    cattr attr = CATTR_DEFAULT;
     for (int j = 0; j < term.cols; j++) {
       termchar *d = chars + j;
       cattr tattr = newchars[j].attr;