+++ src/termclip.c
@@ -1,6 +1,6 @@
 // termclip.c (part of FaTTY)
 // Copyright 2015 Juho Peltonen
-// Based on code from mintty by Andy Koppe and Thomas Wolff
+// Based on code from mintty by 2008-23 Andy Koppe, 2024-2025 Thomas Wolff
 // Adapted from code from PuTTY-0.60 by Simon Tatham and team.
 // Licensed under the terms of the GNU General Public License v3 or later.
 
@@ -513,6 +513,7 @@ void
   term.sel_start = (pos){-sblines(), 0, 0, 0, false};
   term.sel_end = (pos){term_last_nonempty_line(), term.cols, 0, 0, true};
   term.selected = true;
+  term.selection_eq_clipboard = false;
   if (cfg.copy_on_select)
     term_copy();
 }