+++ src/termmouse.c
@@ -1,6 +1,6 @@
 // termmouse.c (part of FaTTY)
 // Copyright 2015 Juho Peltonen
-// Based on code from mintty by 2008-23 Andy Koppe, 2017-24 Thomas Wolff
+// Based on code from mintty by 2008-2023 Andy Koppe, 2017-2025 Thomas Wolff
 // Based on code from PuTTY-0.60 by Simon Tatham and team.
 // Licensed under the terms of the GNU General Public License v3 or later.
 
@@ -269,6 +269,7 @@ static void
   
   //printf("sel_drag %d+%d/2 (anchor %d+%d/2)\n", selpoint.x, selpoint.r, term.sel_anchor.x, term.sel_anchor.r);
   term.selected = true;
+  term.selection_eq_clipboard = false;
   if (!term.sel_rect) {
    /*
     * For normal selection, we set (sel_start,sel_end) to
@@ -653,6 +654,7 @@ bool
         p = get_selpoint(box_pos(p));
         term.mouse_state = MS_OPENING;
         term.selected = true;
+        term.selection_eq_clipboard = false;
         term.sel_rect = false;
         term.sel_start = term.sel_end = term.sel_anchor = p;
         sel_spread();
@@ -678,6 +680,7 @@ bool
       else {
         // Double or triple-click: select whole word or line
         term.selected = true;
+        term.selection_eq_clipboard = false;
         term.sel_rect = false;
         term.sel_start = term.sel_end = term.sel_anchor = p;
         sel_spread();